requirements.txt
AST==0.0.2
attrs==18.2.0
backports.functools-lru-cache==1.5
contextlib2==0.5.5
future==0.16.0
futures==3.2.0
importlib-metadata==0.5
path.py==11.3.0
pathlib2==2.3.2
pluggy==0.7.1
prompt-toolkit==2.0.4
py==1.6.0
Pygments==2.2.0
pyparsing==2.2.0
python-dateutil==2.7.3
PyYAML==3.13
pyyaml-tags==1.0.3
related==0.7.0
scandir==1.9.0
singledispatch==3.4.0.3
six==1.11.0
toml==0.9.6
tox==3.4.0
virtualenv==16.0.0
wcwidth==0.1.7
